requirements.txt
# Telegram Bot Framework
aiogram==3.13.1

# Database (PostgreSQL async)
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0

# Excel file processing
pandas==2.2.2
openpyxl==3.1.5

# Environment variables
python-dotenv==1.0.1

